requirements.txt
django-model-utils==4.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1